(PLATFORM-5172) Remove nCanonicalNamespaces hook
These namespaces are now defined in extension.json already, so there is no need
to redefine them in a hook anymore which was the legacy way, and it prevents us
overridding the namespace list in this hook going forward.

This is a simple change we can also push upstream if we want.